Fixed errors
Fixed the error that a peer is removed and added in a cycle by others peers.

This error originated when a peer disconnected and then tried to connect again on the same port, the other peers still held the dead connection, making it add and remove the re-started peer in a cycle.